Add files via upload
Every session has 10 quizzes and one point for a correct answer (a temporary reward system). The points received in the whole session will be displayed at the end of the game! There's an Exit button, which is supposed to be linked to the Home page. In the views.py code, I've left my debug statements in it with the #DEBUG comments, as David asked. There are just the files required (in my opinion) for merging, but I can upload the whole project if you want to.