Add files via upload
The older version had a bug, where the new locations were marked, but not added to the database. Solved that issue! As discussed, added the feature to let the user add locations by clicking on the map. Added a task constraint, where you can't add more than 2 tasks for a location. Modified some other minor bugs that I found while debugging. 

Note: It's obvious that's a lot of lines for a single file. I will extract the script to a separate .js file once I'm fully done with the app.