Added Features: The map for the user, generation of location graph
For admins: we have 3 files - the map.html (to add new locations, view existing locations, generate a directed graph), delete_location.html (to delete an existing location by selecting), location_form.html (to update or edit an existing location).

For users: just 1 file: user_map.html. This file has a button to generate the graph (as same as in admin's map). Also, the map has the markers added by the admin, but with a different design. 3 colours (red, blue, green) are used to let the user know the status of a particular location. Added also a check-in button which will verify the location of the user.

The 2 Scripts: One is extended on map.html and another one on user_map.html

NOTES: 
- If you're checking GPS, in line 140 and 141 of user_map_script.js, please tweak the value '0.0005' (this value determines the miles). 0.0005 covers a huge radius, so decrease it would be a good starting point. 
- I've used NetworkX for graph generation and Matplotlib for graph designing. If you don't have those libraries installed, use pip.